samurai-examples/testcase/testcase/www/html/testcases/css3/device-adapt/viewport-relative-font-expected.txt

ALERT: viewport size 160x160 scale 2 with limits [2, 5] and userScalable true
